Rebase images onto Centos
Using Alpine linux turned out to be a bad idea once we were using it to
generate compiled.so. Our own versity.xfs and then wheel'ing netifaces
would generate libc symbol messes.

Generating the wheels and .so on Centos and running on Alpine is
similarity messy - glibc isn't available on Alpine and we end up with
missing symbols too.

This is fully tested - building versity.xfs, testing versity.panorama on
mac, building docker for versity.panorama and wheels on mac, running
said bits on a real Linux node.