Add docs for some of our installed packages.
By default, the upstream centos image has a yum.conf with
'tsflags=nodocs' to help keep the image size down. Since this is our
devel image, we need docs to allow easy access to man pages, etc.

We manually set tsflags for some of the installs. If we find holes in
the docs we want, we can roll back the entire yum.conf setting and add
those in. I'm doing a targeted fix now to keep the image size somewhat
sane.

This adds ~25MB to the -devel image, tested with 'man 2 open' and 'man
open_by_handle'.

Do note, you will need to rebuild with the following, as the tsflags
update to the Dockerfile doesn't trigger a full rebuild it seems.

docker build --no-cache -t versity-python:2.7-devel .