Reduce libarchive build requirements
This refactors the list of build requirements down to the minimum needed
to build libarchive RPMs in Docker. We need a few extra packages for
things the system needs to run in creating the source RPM. Finally we
pull down the upstream libarchive src.rpm and make sure we have all of
the BuildRequires installed.